Added a statusBar label and update events to the gui while loading data:
This change adds a label to the buttom of the Qdialog, since this isn't
a main window we can't add a statusBar to it, but we can simulate one.

Using this status bar we can force the gui to update as events are
changing showing the users that things are going on, such as parsing
the metadata and loading the lists. This will allow any message to
be sent duing the packing installing/removing process.